moved some func to test_util
